- Fix warning.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@21082 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/readers/seqdb/seqdbisam.cpp
@@ -160,7 +160,7 @@ CSeqDBIsam::x_SearchIndexNumeric(int              Number,
     Int4 Start     (0);
     Int4 Stop      (m_NumSamples - 1);
     
-    int obj_size = sizeof(SNumericKeyData);
+    int obj_size = (int) sizeof(SNumericKeyData);
     
     while(Stop >= Start) {
         SampleNum = ((Uint4)(Stop + Start)) >> 1;
